1125 version 5.2 - update..!

+++ src/time.c
@@ -6,16 +6,15 @@ void printTime(void){
 
     char timer[10] = "";
     end_time = level_time - (long long int)(cur_time - start_time);
-    if (end_time != ori_time)
+    if (end_time != (double)ori_time)
         npc_move();
     int min = end_time / 60;
     int sec = end_time % 60;
     strcat(timer,ft_itoa(min));
     strcat(timer," : ");
     strcat(timer,ft_itoa(sec));
-
     mvaddstr(timerFrameY+2,timerFrameX+3,timer);
-} 
+}
 // draw 부분으로 이동
 
 // 시간 확인은 end_time 사용
